Split loadout out of loadout row.
