+++ src/app/loadout/Loadout.m.scss
@@ -1,0 +1,99 @@
+@import '../variables.scss';
+
+.loadout {
+  composes: flexColumn from '../dim-ui/common.m.scss';
+  margin-bottom: 10px;
+  padding: 20px;
+  background: rgba(0, 0, 0, 0.25);
+}
+
+.title {
+  composes: flexRow from '../dim-ui/common.m.scss';
+  align-items: center;
+
+  h2 {
+    flex: 1;
+    display: block;
+    margin: 0 !important;
+  }
+
+  @include phone-portrait {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+}
+
+.actions {
+  composes: flexRow from '../dim-ui/common.m.scss';
+  @include horizontal-space-children(4px);
+  @include phone-portrait {
+    margin-top: 16px;
+    margin-bottom: 16px;
+  }
+}
+
+.placeholder {
+  background: rgba(255, 255, 255, 0.05);
+  color: rgba(255, 255, 255, 0.5);
+  width: 100%;
+  height: var(--item-size);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.missingItems {
+  text-transform: none;
+  font-size: 12px;
+  letter-spacing: normal;
+  margin-left: 1em;
+  color: #ccc;
+
+  :global(.app-icon) {
+    margin-right: 0.5em;
+  }
+}
+
+.contents {
+  composes: flexRow from '../dim-ui/common.m.scss';
+  gap: calc(4 * var(--item-margin));
+  flex-wrap: wrap;
+  margin-top: 10px;
+}
+
+.classIcon {
+  width: 16px !important;
+  height: 16px !important;
+  margin-right: 8px;
+}
+
+.mods {
+  flex: 1;
+}
+
+.modsGrid {
+  --item-icon-size: calc(0.8 * var(--item-size));
+  min-width: calc(5 * var(--item-icon-size) + 4 * var(--item-margin));
+  display: grid;
+  grid-template-columns: repeat(auto-fill, var(--item-icon-size));
+  gap: 4px;
+  align-content: flex-start;
+  align-items: flex-start;
+  :global(.item) {
+    --item-size: var(--item-icon-size);
+  }
+}
+
+.showModPlacementButton {
+  margin-top: 8px;
+}
+
+.modsPlaceholder {
+  composes: placeholder;
+  flex: 1;
+}
+
+.loadoutNotes {
+  margin: 4px 0;
+  white-space: pre-wrap;
+}
