+++ src/app/loadout/Loadout.m.scss.d.ts
@@ -1,0 +1,18 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'actions': string;
+  'classIcon': string;
+  'contents': string;
+  'loadout': string;
+  'loadoutNotes': string;
+  'missingItems': string;
+  'mods': string;
+  'modsGrid': string;
+  'modsPlaceholder': string;
+  'placeholder': string;
+  'showModPlacementButton': string;
+  'title': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
