+++ src/app/loadout/Loadouts.m.scss.d.ts
@@ -1,22 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'actions': string;
-  'classIcon': string;
-  'contents': string;
-  'loadout': string;
-  'loadoutNotes': string;
   'menu': string;
   'menuButton': string;
   'menuButtons': string;
-  'missingItems': string;
-  'mods': string;
-  'modsGrid': string;
-  'modsPlaceholder': string;
   'page': string;
-  'placeholder': string;
-  'showModPlacementButton': string;
-  'title': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
